requirements.txt
pyformlang==1.0.10
lark==1.1.9
tqdm==4.66.4
pydantic==2.8.2
networkx==3.3
numpy==1.26.4
pyyaml
torch
transformers
trl
accelerate
bitsandbytes
datasets
